# momentum-core/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The momentum core api manages the core structure of momentum",
        "title": "Momentum Core API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "early-alpha"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/beta/applications": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "artefacts"
                ],
                "summary": "gets a list of all applications",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/artefacts.Artefact"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    }
                }
            }
        },
        "/api/beta/artefact/{id}/": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "artefacts"
                ],
                "summary": "get an artefact by id (an artefact can be any of APPLICATION, STAGE or DEPLOYMENT)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "artefact id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/artefacts.Artefact"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    }
                }
            }
        },
        "/api/beta/deployments": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "artefacts"
                ],
                "summary": "get a list of deployments for a given stage by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "stage id",
                        "name": "stageId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/artefacts.Artefact"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    }
                }
            }
        },
        "/api/beta/dir/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "gets the content of a file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "file id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/files.Dir"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    }
                }
            }
        },
        "/api/beta/file/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "gets the content of a file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "file id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/files.File"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    }
                }
            }
        },
        "/api/beta/file/{id}/line/{lineNumber}/overwrites": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "gets a list of child properties, which are overwritten by the given line.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "file id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "line number in file",
                        "name": "lineNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/files.Overwrite"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    }
                }
            }
        },
        "/api/beta/file/{id}/line/{lineNumber}/overwritten-by": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "gets a list of properties which overwrite the given line.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "file id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "line number in file",
                        "name": "lineNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/files.Overwrite"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    }
                }
            }
        },
        "/api/beta/stages": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "artefacts"
                ],
                "summary": "gets a list of all stages within an application or stage by id.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "application or stage id",
                        "name": "parentId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/artefacts.Artefact"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/config.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "artefacts.Artefact": {
            "type": "object",
            "properties": {
                "contentIds": {
                    "description": "id's of children artefacts",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "description": "id's of parent",
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/artefacts.ArtefactType"
                }
            }
        },
        "artefacts.ArtefactType": {
            "type": "integer",
            "enum": [
                1,
                2,
                4,
                8,
                16,
                32
            ],
            "x-enum-varnames": [
                "ROOT",
                "META",
                "APPLICATION",
                "STAGE",
                "DEPLOYMENT",
                "FILE"
            ]
        },
        "config.ApiError": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "error": {},
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "files.Dir": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/files.File"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "subDirs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/files.Dir"
                    }
                }
            }
        },
        "files.File": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "files.Overwrite": {
            "type": "object",
            "properties": {
                "originFileId": {
                    "type": "string"
                },
                "originFileLine": {
                    "type": "integer"
                },
                "originFilePath": {
                    "type": "string"
                },
                "overwriteFileId": {
                    "type": "string"
                },
                "overwriteFileLine": {
                    "type": "integer"
                },
                "overwriteFilePath": {
                    "type": "string"
                }
            }
        },
...
    }
}